refactor: Car -> domain 폴더로 이동
